add support for setting view by name and slug
