+++ app/packages/core/src/hooks/index.ts
@@ -1,0 +1,1 @@
+export { default as useRefetchableSavedViews } from "./useRefetchableSavedViews";
