+++ app/packages/core/src/index.ts
@@ -1,2 +1,3 @@
 export * from "./components";
 export * from "./plugins";
+export * from "./hooks";
