+++ app/packages/operators/src/built-in-operators.ts
@@ -20,6 +20,8 @@ import {
 } from "./operators";
 import { useShowOperatorIO } from "./state";
 import { useSetRecoilState } from "recoil";
+import useRefetchableSavedViews from "../../core/src/hooks/useRefetchableSavedViews";
+import { toSlug } from "@fiftyone/utilities";
 
 //
 // BUILT-IN OPERATORS
@@ -458,7 +460,10 @@ class SetView extends Operator {
     });
   }
   useHooks(ctx: ExecutionContext): {} {
+    const refetchableSavedViews = useRefetchableSavedViews();
+
     return {
+      refetchableSavedViews,
       setView: fos.useSetView(),
       setViewName: useSetRecoilState(fos.viewName),
     };
@@ -466,14 +471,25 @@ class SetView extends Operator {
   async resolveInput(ctx: ExecutionContext): Promise<types.Property> {
     const inputs = new types.Object();
     inputs.obj("view", { view: new types.HiddenView({}) });
-    inputs.str("name", { label: "Name of a saved view" });
+    inputs.str("name", { label: "Name or slug of a saved view" });
     return new types.Property(inputs);
   }
   async execute({ hooks, params }: ExecutionContext) {
-    if (params.view) {
-      hooks.setView(params.view);
-    } else if (params.name) {
-      hooks.setViewName(params.name);
+    const { view, name } = params || {};
+    if (view) {
+      hooks.setView(view);
+    } else if (name) {
+      const slug = toSlug(name);
+      const savedViews = hooks.refetchableSavedViews?.[0]?.savedViews;
+      const savedView =
+        Array.isArray(savedViews) &&
+        savedViews.find((view) => slug === view.slug);
+      if (!savedView) {
+        throw new Error(
+          `Saved view with name or slug "${name}" does not exist`
+        );
+      }
+      hooks.setViewName(slug);
     } else {
       throw new Error('Param "view" or "name" is required to set a view');
     }
